Change zoom system.

+++ src/Qumulus/Gui/Widgets/EditorView.h
@@ -19,7 +19,7 @@ public:
     EditorView(QWidget* parent = 0);
 
 public slots:
-    void zoom(int value);
+    void zoom(double value);
 
 private:
     QGraphicsScene* mScene;
